refactor + added lombok

+++ src/main/java/com/hse/mappers/EventMapper.java
@@ -16,7 +16,7 @@ public class EventMapper implements RowMapper<Event> {
         event.setId(resultSet.getLong("id"));
         event.setName(resultSet.getString("name"));
         event.setDescription(resultSet.getString("description"));
-        event.setRating(resultSet.getDouble("rating"));
+        event.setRating(resultSet.getFloat("rating"));
         event.setGeoData(resultSet.getString("geoData"));
         event.setSpecialization(Specialization.valueOf(resultSet.getString("specialization")));
         event.setDate(resultSet.getTimestamp("date"));
